requirements.txt
gunicorn==23.0.0
Django==5.1.4
django-environ==0.11.2
scrapy==2.12.0
djangorestframework==3.15.2
celery==5.4.0
redis==5.2.1
django-celery-beat==2.7.0
django_celery_results==2.5.1
drf_spectacular==0.28.0
drf_spectacular_sidecar==2024.12.1
psycopg2-binary==2.9.10
html2text==2024.2.26
scrapy-splash==0.9.0
django-minio-backend==3.7.2
djangorestframework-simplejwt==5.3.1
django_cors_headers==4.6.0
openai==1.57.4
watercrawl-plugin==0.0.2
watercrawl-openai==0.0.1
